employee-api-app/app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
python-multipart==0.0.6
email-validator==2.1.0
python-jose[cryptography]==3.3.0
google-cloud-secret-manager==2.16.4
